Move component.ts and input.ts into modules
