+++ src/components/cancel-save/cancel-save.ts
@@ -2,7 +2,7 @@
 
 import './cancel-save\.scss';
 
-import { Component } from '../../component';
+import { Component } from '../../modules/component';
 import { Button, ClickCallback } from '../button/button';
 
 interface CancelSaveProps {
