+++ src/components/input-file/input-file.ts
@@ -2,7 +2,7 @@
 
 import './input-file\.scss';
 
-import { Input } from '../../input';
+import { Input } from '../../modules/input';
 import { Button } from '../button/button';
 
 /**
