+++ src/components/link-inner/link-inner.ts
@@ -2,7 +2,7 @@
 
 import './link-inner\.scss';
 
-import { Component } from '../../component';
+import { Component } from '../../modules/component';
 import { loadPath } from '../../modules/router';
 
 /**
