+++ src/components/page-profile/page-profile.ts
@@ -2,7 +2,7 @@
 
 import './page-profile\.scss';
 
-import { Component } from '../../component';
+import { Component } from '../../modules/component';
 import { FormProfilePublic } from '../form-profile-public/form-profile-public';
 import { WalletOptions } from '../wallet-options/wallet-options';
 import { InputField } from '../input-field/input-field';
