+++ src/components/page-sign-up/page-sign-up.ts
@@ -1,6 +1,6 @@
 'use strict';
 
-import { Component } from '../../component';
+import { Component } from '../../modules/component';
 import { loadPath } from '../../modules/router';
 import { Button } from '../button/button';
 import { FormSignUp } from '../form-sign-up/form-sign-up';
