+++ src/components/wallet-options/wallet-options.ts
@@ -2,7 +2,7 @@
 
 import './wallet-options\.scss';
 
-import { Component } from '../../component';
+import { Component } from '../../modules/component';
 import { Button } from '../button/button';
 import { InputField } from '../input-field/input-field';
 import { topUpAmountGetError } from '../../modules/validation';
