+++ src/modules/validation.ts
@@ -1,6 +1,6 @@
 'use strict';
 
-import { Input } from '../input';
+import { Input } from './input';
 
 /**
  * Проверка валидности email
